+++ CHANGELOG.md
@@ -4,6 +4,12 @@
 
 ## History
 
+### current: 0.2.0 (2022 Aug xx)
+
+ - new: `Timestamp` wrapper around timestamps
+ - update: `Timespan`: get durations with specified units only
+ - update: `Timespan`: add month unit
+
 ### 0.1.0 (2022 Aug 13)
 
  - initial